add text fields in ColorSliderView
