+++ ColorSliders/TextField+Extensions.swift
@@ -1,0 +1,28 @@
+//
+//  TextField+Extensions.swift
+//  ColorSliders
+//
+//  Created by Михаил Зиновьев on 25.01.2022.
+//
+
+import SwiftUI
+
+struct BorderedViewModifier: ViewModifier {
+    func body(content: Content) -> some View {
+        content
+            .padding(EdgeInsets(top: 8, leading: 16, bottom: 8, trailing: 16))
+            .overlay(
+                RoundedRectangle(cornerRadius: 4)
+                    .stroke(lineWidth: 2)
+                    .foregroundColor(.blue)
+            )
+            .shadow(color: Color.gray.opacity(0.4), radius: 3, x: 1, y: 2)
+    }
+}
+
+extension TextField {
+    func bordered() -> some View {
+        ModifiedContent(content: self, modifier: BorderedViewModifier())
+    }
+}
+
